+++ frontend/components/community/index.ts
@@ -1,7 +1,7 @@
 import KeywordAdder from './KeywordAdder';
 import CommunityHeader from './CommunityHeader';
 import CommunityLayout from './CommunityLayout';
-import KeywordAddModal from './KeywordAddModal';
+import KeywordAddModalContent from './KeywordAddModalContent';
 import KeywordBubbleChart from './KeywordBubbleChart';
 import MyKeyword from './MyKeyword';
 import MyKeywordList from './MyKeywordList';
@@ -11,7 +11,7 @@ import AutoCompleteFormLayout from './AutoCompleteFormLayout';
 export {
   CommunityHeader,
   CommunityLayout,
-  KeywordAddModal,
+  KeywordAddModalContent,
   KeywordBubbleChart,
   MyKeyword,
   MyKeywordList,
